PrivacyGuardian_final_zip/PrivacyGuardian/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "flask>=3.1.2",
    "matplotlib>=3.10.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "python-dateutil>=2.9.0.post0",
    "reportlab>=4.4.4",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
]
